Civil Liberties and Political Rights Ratings, 1972-2023

For each country and territory, Freedom in the World analyzes the electoral process, political pluralism and participation, the functioning of the government, freedom of expression and of belief, associational and organizational rights, the rule of law, and personal autonomy and individual rights.

country_rating_statuses

Format

A data frame with 9,238 observations and 9 variables.

Source

Adapted from Freedom House.

Variables

year: Year of observation (1972-2023). The survey is applied the year after.

country: Country name.

iso2c: ISO 2-character country code. Czechoslovakia, Kosovo, Micronesia, Serbia and Montenegro, and Yugoslavia do not have unambiguous matches and appear as 'NA'.

iso3c: ISO 3-character country code. Same criteria for 'NA' as in iso2c.

continent: Continent name.

year: Year of observation (1973-2023).

political_rights: Political rights rating (1-7 scale, with one representing the highest degree of Freedom and seven the lowest).

civil_liberties: Civil liberties rating (1-7 scale, with one representing the highest degree of Freedom and seven the lowest).

status: Status of the country (Free, Partly Free, Not Free).

color: Color associated with the status of the country.

References

Freedom House. 2024. Freedom in the World 2024: The Mounting Damage of Flawed Elections and Armed Conflict. Washington, DC: Freedom House. https://freedomhouse.org/report/freedom-world/2024/mounting-damage-flawed-elections-and-armed-conflict. Accessed 2024-04-25.